:sparkles: feat : 주문 기능 구현

+++ src/test/java/kr/hhplus/be/server/stub/UserFixture.kt
@@ -6,7 +6,7 @@ object UserFixture {
     fun create(
         name: String = "유저 A",
         balance: Long = 0,
-        version: Int = 0,
+        version: Long = 0,
         id: Long = 1L,
     ): User = User(name = name, balance = balance, version = version, id = id)
 }
